PrimeGUI1/src/app/app.component.spec.ts
import { NO_ERRORS_SCHEMA } from '@angular/core';
import { async, TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';
import { AppComponent } from './app.component';
import { HomeComponentComponent } from './home-component/home-component.component';
import { ProxyComponentComponent } from './proxy-component/proxy-component.component';
import { TaskComponentComponent } from './task-component/task-component.component';

describe('AppComponent', () => {
  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [
        RouterTestingModule
      ],
      declarations: [
        AppComponent
      ],
    }).compileComponents();
  });

  it('should create the app', () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });

  it(`should have as title 'PrimeGUI1'`, () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app.title).toEqual('PrimeGUI1');
  });

  it('should render title', () => {
    const fixture = TestBed.createComponent(AppComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('.content span')?.textContent).toContain('PrimeGUI1 app is running!');
  });
});

beforeEach(async(() => {
  TestBed.configureTestingModule({
    declarations: [ TaskComponentComponent,ProxyComponentComponent,HomeComponentComponent ],
    schemas: [NO_ERRORS_SCHEMA]
  })
  .compileComponents();
}));
